\name{RSI data}
\alias{retail}
\docType{data}
\title{
Retail Sales Index (RSI) data
}
\description{
This data frame is taken from the online data resource of the
U.K.'s Office for National Statistics. It contains time-indexed values
of a retail sales index, a figure describing the turnover of retail
businesses as a percentage of a 2010 baseline.
}
\details{
The original data files are no longer hosted on the main ONS webpages but, for the foreseeable future, ought to be accessible via the UK's National Archives \url{https://webarchive.nationalarchives.gov.uk/ukgwa/20160105160709/http://www.ons.gov.uk/ons/index.html}.
}
\references{
Nason, G.P., Powell, B., Elliott, D. and Smith, P. (2016)
	Should We Sample a Time Series More Frequently? Decision Support
	via Multirate Spectrum Estimation.
	Journal of the Royal Statistical Society, Series A., 179,
	(to appear).
}
\keyword{datasets}
